[LLC] Do not unconditionally emit a body parameter (#1129)
We previously were always adding a `RequestContent bodyContent`
parameter for all our operation methods in the low level
client. However we shouldn't do that if the operation didn't have any
body parameters as part of its definiton (one common example, `GET`
operations, which do not take a request body).

While I was in the area, I saw that there was a little bit of
simplification we could make to the generator if we were willing to
slightly change the generated code for the LLC such that we always
have both a message and a request. This also means that for LLC
methods we'll now also correctly mark operations which should not
buffer their response.