[LLC] Wrap service calls in a diagnotics scope
Introduce a `LowLevelClientMethod` which is sort of like the DataPlane
`ClientMethod` (with a little less state). It holds the underlying
`RestClientMethod` that we use for the service call, as well a new
`Diagnostic` so we can have the correct name for our scope.

Much of this logic is similar to what the dataplane client is doing
today, hopefully we can unify this down the line, but for now this
unblocks test recording for.

Fixes #1214