[LLC] Use `Azure.Core.ContentType` for content type
When an operation consumes multiple content types, the generated
operation method includes a parameter for the content type. For low
level clients, we'd like to use the new `Azure.Core.ContentType` type
for this parameter. In addition, we'd like it to come /after/ the
parameter for the body of the operation. This allows us to add helpers
where the content type is infered while keeping the content in the
same location.

For example:

```csharp
// Hand authored overload that calls SendMessage(..., ContentType.TextPlain);
Response SendMessage(string s);
// Hand authored overload that calls SendMessage(..., ContentType.ApplicationOctetStream);
Response SendMessage(Stream s);
// Generated protocol method.
Response SendMessage(RequestContent content, ContentType contentType);
```

Fixes #1333