Update StyleCop to Latest
The existing version of StyleCop we use did not understand records
(and the "File Name should be the same as Type Name rule blows up if
it sees one) and we want to change `Parameter` to be a record as part
of #1336.

This upgrades StyleCop to the latest pre-release version on NuGet and
fixes the new warnings it flagged (ensuring newlines at the end of
some files and writing `new(` instead of `new (`.