refactor(generator): remove `int16` type
`int16` (or `short`) type is actually not supported in either swagger spec or in autorest.
- For swagger, see: https://docs.swagger.io/spec.html#431-primitives
- For autorest:
  - The supported integer formats are here: https://github.com/Azure/autorest/blob/7c68312403154a29770a1783fdca152a67fcc180/packages/libs/openapi/src/formats.ts#L51
  - In the modelfour test case, `int16` will be transformed to `int32`: https://github.com/Azure/autorest/blob/7c68312403154a29770a1783fdca152a67fcc180/packages/extensions/modelerfour/test/modeler/modelerfour.test.ts#L179

Besides, if we define `int16` or `short` as the `format`, autorest will print the warning messages like below:
```
WARNING (Modeler): Integer schema 'schemas:84' with unknown format:
'short' is not valid.  Treating it as 'int32'.
```

So we'd better remove it, to align with upstream specs.