[LLC] Have `LowLevelRestClient` extend `RestClient`
Refactor `LowLevelRestClient` to be more like `DataPlaneRestClient`,
it now reuses the `RestClient` base class.

As part of this, we started to generate multiple methods for an
operation with multiple service requests due to divergent logic. That
exposed the same issue we outlined in #1215. I have removed the
documentation from these internal methods to address this issue (they
had doc comments because at one point they were publicly exposed
methods and we did not remove them when we decided to make the methods
private).