[LLC] Model LROs as `Operation<BinaryData>` (#1384)
For long running operations, we want to use the `Operation<T>`
abstraction to help users deal with long running operations. Since we
do not have model types, we say the result of a long running operation
for a protocol method is just the contents of the body of the final
state of the operation.

This change introduces `LowLevelBinaryDataOperation`, an internal
type, which is used by all protocol methods. This type in not publicly
exposed, instead we just type LROs as returning instances of
`Operation<BinaryData>`, which this internal type implements.

Fixes #1328