[LLC] Refactoring to share more code with DataPlane (#1430)
* [LLC] Have `LowLevelRestClient` extend `RestClient`

Refactor `LowLevelRestClient` to be more like `DataPlaneRestClient`,
it now reuses the `RestClient` base class.

As part of this, we started to generate multiple methods for an
operation with multiple service requests due to divergent logic. That
exposed the same issue we outlined in #1215. I have removed the
documentation from these internal methods to address this issue (they
had doc comments because at one point they were publicly exposed
methods and we did not remove them when we decided to make the methods
private).

* [LLC] Update Generation

* [LLC] Refactor how `RequestOptions` parameter works

Stop putting a `RequestOptions` parameter in the Parameters list for
the method when we model it, instead add the parameter when we are
writing code for the method.

This mirrors what DataPlane does and removes another special case from
`LowLevelRestClient`.

Fixes #1222

* [LLC] Elide documentation on internal methods

The `Create<OperationName>Request` methods were being documented
because at one point they were public. This is no longer the case, so
we remove the documentation since it could be ambiguous at times
(since the specific overload parameters are not included in the cref)

Fixes #1215

* [LLC] Unify LowLevel and DataPlane `WriteRequestCreation`

There are a few cases in `WriteRequestCreation` where we took into
account if we were generating code for an LLC or a DataPlane client,
which we can unify by ensuring that the generated LLC client has a
`_pipeline` field, which is exposed by the `Pipeline` property.

* [LLC] Remove unused constant

* [LLC] Refactor how RequestOptions parameter is emitted

* [LLC] Use Append instead of Concat