[LLC] Split DataPlane and Rest Clients
To prepare for the pageable changes (which were more difficult to
implement with the current architecture because a single operation
could end up using multiple different REST operations) we adopt a
similar client factoring to what we have in the Track 2 Data Plane
client generators. We now have an internal
`<OperationGroupName>RestClient` which deals with the actual REST
layer and a `<OperationGroupName>Client` client type, which exposes
these REST client methods (and in the case of LROs and Pageables
exposes them as `Operation<T>/AsyncPageable<T>`).

Having a similar split between public client and REST client should
also make it easier to leverage existing Track 2 High Level Client
behaviors.