[LLC] Support Pageable Operations
This change adds support for returning `Pageable<T>/AsyncPageable<T>`
from operations which return paged data. Since we do not have models,
each item is a `BinaryData` which contains the JSON text of the
underlying item in the collection. These BinaryData objects are formed
**over** the existing data that backs the response (by slicing the
binary data that backs the underlying response).

In addition to just normal pagable operations, some of the code for
long running operations had to be updated to handle the fact that the
final state of an operation may itself be a pageable collection (in
which case we return `Operation<Pageable<BinaryData>>` instead of just
`Operation<BinaryData>`.

Fixes #1329