[LLC] Disable Flatting and Grouping for LLC
Flatting and Grouping are two autorest features which cause parameters
to move around from their write locations. Since LLC's goal is to is
to not do these sort of transforms, we disable them in our autorest
configuration.

Since the generated code wasn't using this information, none of the
actual code changes. However, we were using schemas for the operations
(which had already been flatted/grouped) for generating docmentation,
so there are some slight tweeks to the generated docs (and they now
correctly reflect the service payload).

Fixes #1316