[LLC] Support Generating "Sub Clients"
Previously, the code generator would generate a publicly constructable
client type for each operation group. This leads to many top level
clients which can lead to user confusion, because it is not clear
which client is the "entry point" to the service. The Sub Clients
feature allows a way to configure the code generator such that it does
not generate publicly consturctable clients for certian operation
groups and instead generates a factory method on a different client
which can be used to construct the client (which reuses the same
underlying HttpPipeline).

A new sample based on the Purview Accounts service is added which
shows how this can be used to change the factoring of the client
libraries (and allows us to machine generate what was previously
handled via hand authored customizations to the client library)

Fixes #1527