[LLC] Reduce number of Long Running Operation types
As part of introducing support for LRO's which produce pageable
responses, we introduced a new `FuncOperation` type and then
implemented LowLevelBinaryDataOperation in terms of it via
inheritance. After talking with Alexander, we'd rather not have two
types here.

I've removed `LowLevelBinaryDataOperation` in favor of just using
`LowLevelFuncOperation` with a static selector.