[LLC] Pageable Support (#1514)
* [LLC] Split DataPlane and Rest Clients

To prepare for the pageable changes (which were more difficult to
implement with the current architecture because a single operation
could end up using multiple different REST operations) we adopt a
similar client factoring to what we have in the Track 2 Data Plane
client generators. We now have an internal
`<OperationGroupName>RestClient` which deals with the actual REST
layer and a `<OperationGroupName>Client` client type, which exposes
these REST client methods (and in the case of LROs and Pageables
exposes them as `Operation<T>/AsyncPageable<T>`).

Having a similar split between public client and REST client should
also make it easier to leverage existing Track 2 High Level Client
behaviors.

* [LLC] Support Pageable Operations

This change adds support for returning `Pageable<T>/AsyncPageable<T>`
from operations which return paged data. Since we do not have models,
each item is a `BinaryData` which contains the JSON text of the
underlying item in the collection. These BinaryData objects are formed
**over** the existing data that backs the response (by slicing the
binary data that backs the underlying response).

In addition to just normal pagable operations, some of the code for
long running operations had to be updated to handle the fact that the
final state of an operation may itself be a pageable collection (in
which case we return `Operation<Pageable<BinaryData>>` instead of just
`Operation<BinaryData>`.

Fixes #1329

* [LLC] Make LowLevelPageableHelpers.GetItemsAndNextLinkFromJson private

* [LLC] Small tweaks to generated code

- Don't keep client level parameters as fields in the DataPlane client.
- Use `_restClient` as the field name for the REST Client, not `RestClient`.

* [LLC] Use more specific names for some state

* [LLC] Update Generated Test Projects

* [LLC] Call `ToArray` on `IEnumerable` parameters for Pageables

Doing so ensures that we only enumerate the parameter to the client
method once, even if we end up calling multiple service methods to
fetch multiple pages.

* [LLC] Small fixes from PR Feedback

* [LLC] Reduce number of Long Running Operation types

As part of introducing support for LRO's which produce pageable
responses, we introduced a new `FuncOperation` type and then
implemented LowLevelBinaryDataOperation in terms of it via
inheritance. After talking with Alexander, we'd rather not have two
types here.

I've removed `LowLevelBinaryDataOperation` in favor of just using
`LowLevelFuncOperation` with a static selector.