fix(test): date parsing error in some time zones
We have some test cases which test if `0001-01-01` is within valid date time range. The problem is that if you parse `0001-01-01` as `DatetimeOffset`, it will assume local time zone, and that will result in error in some time zones. Those errors will not happen in CI pipline of which the server is set under UTC+0. They probably won't show up in US timezone either.
The fix is simple, just specifying the time zone in the test data.