feat(generator): convert datetimeoffset to utc time zone before serializing to string
In track1 mgmt SDK, we use `DateTime` which is always serialized into `YYYY-MM-DDZ` format.
In track2, we use `DateTimeOffset` which is serialized to various formats, according to the time zone in `DateTimeOffset` instance. That will cause some backend server errors, since they do not accept format other than `YYYY-MM-DDZ`.
This commit is a conservative fix. It will change the generated serialization codes to always convert timezone to UTC+0, before serializing it to string.