fix(common): datetime header value with no timezone should assume utc+0
We have unit test failure in machines with UTC+8 timezone (actually any timezone with positive offset), which is due to the parsing logic of header values. It assume local timezone when parsing datetime header values. When parsing datetime header value with no timezone, we should assme utc+0 which is a best practice.

This commit will:
1. Replace the parsing logic in `ResponseHeadersExtensions` with the common logic in `TypeFormatter.ParseDateTimeOffset` which will take care of default timezone.
2. Update unit test cases . Finalize the timezone in the expected datetime value.

fix #1768