refactor(mgmt): rename `UserAgentOverride` to `SDKUserAgent` (#1930)
In https://github.com/Azure/azure-sdk-for-net/pull/26556, we've decided to rename `UserAgentOverride` to `SDKUserAgent`, because now we allow users to append `User-Agent` headers. Accordingly, we should update the property name in generator. This change should be limited to control plane only, because `RestClientWriter.UseUserAgentOverride` is `false` by default, which is overridden in `MgmtRestClientWriter.UseUserAgentOverride` to return `true`.

This commit:
1. Rename property `UserAgentOverride` to `SDKUserAgent`
2. Update the samples and test projects

resolve https://github.com/Azure/azure-sdk-for-net/issues/26406