fix(common): random error due to repetitive loading configuration
We have mulitple `IPlugin` which could load configuration multiple times.
For example, in `src/AutoRest.CSharp/readme.md`, we load plugins `csharpgen` and `csharpproj`.
The bad thing is that plugin invocation is async, so loading configuration could be concurrent. And unfortunately the configuration loading is not concurrency safe. For example, `_intrinsicTypesToTreatEmptyStringAsNull.UnionWith(additionalIntrinsicTypesToTreatEmptyStringAsNull);` is not thread-safe.

To fix that, we should make sure the configuration is loaded only once. This commit uses a flag `IsInitialized` and a lock `InitLock` to make sure the configuration initialization is executed only once and exclusively.

fix #3697