feat(dpg): generate polling operation by default (#3981)
* feat(dpg): generate polling operation by default

Polling operation can be both public and internal. If we by default do not
generate polling operations, then there is no way for developers to generate
public polling operations. We can use `@access` to hide polling operations, or
define LRO in other ways so that polling operation will not be defined.

- change emitter to emit polling operation by default
- regen test/samples
- add test case for lro polling operation in namespace without @service

resolve #3964

---------

Co-authored-by: Mingzhe Huang (from Dev Box) <mingzhehuang@microsoft.com>